requirements.txt
neo4j
python-dotenv
python-bitcoinrpc
requests
numpy
matplotlib
networkx
plotly
pyvis
seaborn
scikit-learn
scipy